pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
pyjwt = "^2.10.1"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
sqlalchemy = "^2.0.36"
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
pandas = "^2.2.3"
selenium = "^4.27.1"
seaborn = "^0.13.2"
scipy = "^1.14.1"
matplotlib = "^3.10.0"
mplfinance = "^0.12.10b0"
scikit-learn = "^1.6.0"
xgboost = "^2.1.3"
tensorflow = "^2.18.0"
statsmodels = "^0.14.4"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
